feat: add the global websoccket
