+++ client/components/Chats.tsx
@@ -5,37 +5,10 @@ import {
 import { useAuth0 } from '@auth0/auth0-react'
 import { useParams, useNavigate } from 'react-router'
 import Message from './Message.tsx'
-import { useQueryClient } from '@tanstack/react-query'
 import { useEffect } from 'react'
 import { useUsers } from '../hooks/useUsers.ts'
 
 function Chat() {
-  const queryClient = useQueryClient()
-
-  useEffect(() => {
-    const ws = new WebSocket('ws://localhost:3000/ws')
-
-    ws.onopen = () => {
-      console.log('websocket connected')
-    }
-
-    ws.onmessage = (event) => {
-      const data = JSON.parse(event.data)
-      if (data.type === 'database_change') {
-        queryClient.invalidateQueries({ queryKey: ['chatsById'] })
-        queryClient.invalidateQueries({ queryKey: ['chats'] })
-      }
-    }
-
-    ws.onclose = () => {
-      console.log('Websocket closed')
-    }
-
-    ws.onerror = (error) => {
-      console.log(error)
-    }
-  }, [queryClient])
-
   const { id } = useParams<{ id: string }>()
   const chatId = Number(id)
   const { chatById, messageByChatId } = useGetMessageByChatId(chatId)
